pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edraft"
version = "0.1.0"
description = "Local Outlook reply-draft assistant that reads unread mail and creates draft replies without sending."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "edraft" }
]
dependencies = [
  "beautifulsoup4>=4.12.3",
  "httpx>=0.27.2",
  "msal>=1.31.1",
  "openai>=1.54.0",
  "python-dotenv>=1.0.1",
  "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[project.scripts]
edraft = "edraft.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
